Change byte order to be big endian
Network byte order, and the IETF in general, is big endian,
while Intel went with little endian CPU.
As binary serialization can be used mostly for disk serialization
and network communication, it doesn't make much difference in the first
case, and makes a whole lot of difference in the second.
Instead of allowing the user to choose the byte order,
as suggested in issue 8, we simply change it to big endian by default,
as there is currently no foreseen use case for little endian.