Introduce deserializeArray for length-less records
When implementing DNS RR records, the length of the record is provided
in a separate field, and the data is not prefixed by the length,
so we need to implement a manual fromBinary. But it is an error-prone endeavour,
as it needs to duplicate the string validation to be correct.
deserializeArray is a simpler, and correct way to do it.